feat: transfer drafts finalize at creation — Safe-wallet visibility restored
Queued transfer requests were invisible in app.safe.global until approved:
the lazy-draft design deferred nonce/hash/mirror to approval time for every
kind. Users expect the pre-draft behavior — agent signs at creation and the
tx sits in the Safe queue at 1/2 immediately.

Finalization is now per-kind, via a shared finalizeDraft() used by both
agentPropose and POST /transactions/:id/finalize:

- TRANSFERS finalize eagerly at creation — their calldata can't go stale.
  Dismissing one parks its nonce until superseded, as the pre-draft flow
  always did.
- SWAPS stay lazy — finalize re-quotes route + min-out when the user
  approves (a queue-time quote reverts on-chain by signing time), so a
  dismissed swap draft never parks a nonce.
